+++ src/CoreNutrition.Infrastructure/Security/TokenGenerator/TokenValidationConfiguration.cs
@@ -1,0 +1,1 @@
+namespace CoreNutrition.Infrastructure.Security.TokenGenerator;
